Refactor: Adding calculations
