+++ future-apps/payroll/app/src/screens/components/AvailableSalary.js
@@ -3,6 +3,7 @@ import styled from 'styled-components'
 import AvailableSalaryTable from './AvailableSalaryTable'
 import { formatCurrency, SECONDS_IN_A_YEAR } from '../../utils/formatting'
 import { differenceInSeconds } from 'date-fns'
+import { summation } from '../../utils/calculations'
 
 import { connect } from '../../context/AragonContext'
 import Section from '../../components/Layout/Section'
@@ -32,10 +33,9 @@ class AvailableSalary extends React.PureComponent {
   }
 
   sumExchanged (payments, accountAddress) {
-    const init = 0
-    const reducer = (acc, payment) => acc + payment.exchanged
+    const field = 'exchanged'
     const filter = e => e.accountAddress === accountAddress
-    const totalTransferred = payments.filter(filter).reduce(reducer, init)
+    const totalTransferred = summation(payments.filter(filter), field)
     return totalTransferred
   }
